requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.1
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-tenant-schemas==1.10.0
django-tenants==3.5.0

# Authentication & Security
django-allauth==0.57.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# File Processing
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# AI & ML
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
openai==1.3.5
requests==2.31.0

# Data Visualization
bokeh==3.3.2
dash==2.14.2
dash-bootstrap-components==1.5.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# File Storage
django-storages==1.14.2
boto3==1.34.0

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Payment Processing
stripe==7.6.0

# Development
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
